% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aragog}
\alias{aragog}
\title{Get the spider's name}
\usage{
aragog()
}
\value{
string. The spider's name
}
\description{
aragog
}
\details{
Get Hagrid's spider's name
}
\author{
Insights Engineering <ie@example.com>
}
